+++ apps/CMakeLists.txt
@@ -1,0 +1,6 @@
+ADD_SUBDIRECTORY(cmd)
+IF(QT4_FOUND)
+    ADD_SUBDIRECTORY(gui)
+ELSE(QT4_FOUND)
+    MESSAGE(STATUS " NOTE: Qt4 not found. Not building GUI app.")
+ENDIF(QT4_FOUND)
