+++ apps/cmd/CMakeLists.txt
@@ -1,0 +1,6 @@
+INCLUDE_DIRECTORIES(${POKEHACK_SOURCE_DIR}/include)
+
+ADD_EXECUTABLE(pokehack_cmd pkg3.c)
+TARGET_LINK_LIBRARIES(pokehack_cmd pokehack)
+
+INSTALL(TARGETS pokehack_cmd RUNTIME DESTINATION ${RUNTIME_DIR} COMPONENT Runtime)
