--- data/attacks.txt
@@ -1,356 +1,0 @@
-No Attack
-Pound
-Karate Chop
-Double Slap
-Comet Punch
-Mega Punch
-Pay Day
-Fire Punch
-Ice Punch
-Thunderpunch
-Scratch
-Vicegrip
-?
-Razor Wind
-Swords Dance
-Cut
-Gust
-Wing Attack
-Whirlwind
-Fly
-Bind
-Slam
-Vine Whip
-Stomp
-Double Kick
-Mega Kick
-Jump Kick
-Rolling Kick
-Sand Attack
-Headbutt
-Horn Attack
-Fury Attack
-Horn Drill
-Tackle
-Body Slam
-Wrap
-Take Down
-Thrash
-Double-Edge
-Tail Whip
-Poison Sting
-Twineedle
-Pin Missile
-Leer
-Bite
-Growl
-Roar
-Sing
-Supersonic
-Sonicboom
-Disable
-Acid
-Ember
-Flamethrower
-Mist
-Water Gun
-Hydro Pump
-Surf
-Ice Beam
-Blizzard
-Psybeam
-Bubblebeam
-Aurora Beam
-Hyper Beam
-Peack
-Drill Peck
-Submission
-Low Kick
-Counter
-Strength
-Seismic Toss
-Absorb
-Mega Drain
-Leech Seed
-Growth
-Razor Leaf
-Solarbeam
-Poisonpowder
-Stun Spore
-Sleep Powder
-Petal Dance
-String Shot
-Dragon Rage
-Fire Spin
-Thundershock
-Thunderbolt
-Thunder Wave
-Thunder
-Rock Throw
-Earthquake
-Fissure
-Dig
-Toxic
-Confusion
-Psychic
-Hypnosis
-Meditate
-Agility
-Quick Attack
-Rage
-Teleport
-Night Shade
-Mimic
-Screech
-Double Team
-Recover
-Harden
-Minimize
-Smokescreen
-Confuse Ray
-Withdraw
-Defense Curl
-Barrier
-Light Screen
-Haze
-Reflect
-Focus Energy
-Bide
-Metronome
-Mirror Move
-Selfdestrust
-Egg Bomb
-Lick
-Smog
-Sludge
-Bone Club
-Fire Blast
-Waterfall
-Clamp
-Swift
-Skull Bash
-Spike Cannon
-Constrict
-Amnesia
-Kinesis
-Softboild
-Hi Jump Kick
-Glare
-Dream Eater
-Poison Gas
-Barrage
-Leech Life
-Lovely Kiss
-Sky Attack
-Transform
-Bubble
-Dizzy Punch
-Spore
-Flash
-Psywave
-Splash
-Acid Armor
-Crabhammer
-Explosion
-Fury Swipes
-Bonemerang
-Rest
-Rock Slide
-Hyper Fang
-Sharpen
-Conversion
-Tri Attack
-Super Fang
-Slash
-Substiture
-Struggle
-Sketch
-Triple Kick
-Thief
-Spider Web
-Mind Reader
-Nightmare
-Flame Wheel
-Snore
-Curse
-Flail
-Conversion 2
-Aeroblast
-Cotton Spore
-Reversal
-Spite
-Powder Snow
-Protect
-Mach Punch
-Scary Face
-Faint Attack
-Sweet Kiss
-Belly Drum
-Sludge Bomb
-Mud-Slap
-Octazooka
-Spikes
-Zap Cannon
-Foresight
-Destiny Bond
-Perish Song
-Ice Wind
-Detect
-Bone Rush
-Lock-On
-Outrage
-Sandstorm
-Giga Drain
-Endure
-Charm
-Rollout
-False Swipe
-Swagger
-Milk Drink
-Spark
-Fury Cutter
-Steel Wing
-Mean Look
-Attract
-Sleep Talk
-Heal Bell
-Return
-Present
-Frustration
-Safeguard
-Pain Split
-Sacred Fire
-Magnitude
-Dynamicpunch
-Megahorn
-Dragonbreath
-Baton Pass
-Encore
-Pursuit
-Rapid Spin
-Sweet Scent
-Iron Tail
-Metal Claw
-Vital Throw
-Morning Sun
-Synthesis
-Moonlight
-Hidden Power
-Cross Chop
-Twister
-Rain Dance
-Sunny Day
-Crunch
-Mirror Coat
-Psych Up
-Extremespeed
-Ancientpower
-Shadow Ball
-Future Sight
-Rock Smash
-Whirlpool
-Beat Up
-Fake Out
-Uproar
-Stockpile
-Spit Up
-Swallow
-Heat Wave
-Hail
-Torment
-Flatter
-Will-O-Wisp
-Memento
-Facade
-Focus Punch
-Smellingsalt
-Follow Me
-Nature Power
-Charge
-Taunt
-Helping Hand
-Trick
-Role Play
-Wish
-Assist
-Ingrain
-Superpower
-Magic Coat
-Recycle
-Revenge
-Brick Break
-Yawn
-Knock Off
-Endeavor
-Eruption
-Skill Swap
-Imprison
-Refresh
-Grudge
-Snatch
-Secret Power
-Dive
-Arm Thrust
-Camouflage
-Tail Glow
-Luster Purge
-Mist Ball
-Featherdance
-Teeter Dance
-Blaze Kick
-Mud Sport
-Ice Ball
-Needle Arm
-Slack Off
-Hyper Voice
-Poison Fang
-Crush Claw
-Blast Burn
-Hydro Cannon
-Meteor Mash
-Astonish
-Weather Ball
-Aromatherapy
-Fake Tears
-Air Cutter
-Overheat
-Odor Sleuth
-Rock Tomb
-Silver Wind
-Metal Sound
-Grasswhistle
-Tickle
-Cosmic Power
-Water Spout
-Signal Beam
-Shadow Punch
-Extrasensory
-Sky Uppercut
-Sand Tomb
-Sheer Cold
-Muddy Water
-Bullet Seed
-Aerial Ace
-Icecle Spear
-Iron Defense
-Block
-Howl
-Dragon Claw
-Frenzy Plant
-Bulk Up
-Bounce
-Mud Shot
-Poison Tail
-Cover
-Volt Tackle
-Magical Leaf
-Water Sport
-Calm Mind
-Leaf Blade
-Dragon Dance
-Rock Blast
-Shock Wave
-Water Pulse
-Doom Desire
-Psycho Boost
--end-
