--- data/items.txt
@@ -1,227 +1,0 @@
-0 - No Item
-1 - master ball
-2 - ultra ball
-3 - great ball
-4 - poke ball
-5 - safari ball
-6 - net ball
-7 - dive ball
-8 - nest ball
-9 - repeat ball
-10- timer ball
-11 - luxury ball
-12 - premier ball
-13 - potion
-14 - antidote
-15 - burn heal
-16 - ice heal
-17 - awakening
-18 - parlyz heal
-19 - full restore
-20 - max potion
-21 - hyper potion
-22 - super potion
-23 - full heal
-24 - revive
-25 - Max Revive
-26 - Fresh Water
-27 - Soda Pop
-28 - Lemonade
-29 - Moomoo Milk
-30 - Energypowder
-31 - Energy Root
-32 - Heal Powder
-33 - Revival Herb
-34 - Ether
-35 - Max Ether
-36 - Elixir
-37 - Max Elixir
-38 - Lava Cookie
-39 - Blue Flutre
-40 - Yellow Flute
-41 - Red Flute
-42 - Black Flute
-43 - White Flute
-44 - Berry Juice
-45 - Sacred Ash
-46 - Shoal Salt
-47 - Shoal Shell
-48 - Red Shard
-49 - Blue Shard
-50 - Yellow Shard
-51 - Green Shard
-52 - ????????
-53 - ????????
-54 - ????????
-55 - ????????
-56 - ????????
-57 - ????????
-58 - ????????
-59 - ????????
-60 - ????????
-61 - ????????
-62 - ????????
-63 - HP UP
-64 - Protein
-65 - Iron
-66 - Carbos
-67 - Calcium
-68 - Rare Candy
-69 - PP UP
-70 - Zinc
-71 - PP MAX
-72 - ????????
-73 - Guard Spec
-74 - Dire Hit
-75 - X Attack
-76 - X Defend
-77 - X Speed
-78 - X Accuracy
-79 - X Special
-80 - Poke Doll
-81 - Fluffy Tail
-82 - ????????
-83 - Super Repel
-84 - Max Repel
-85 - Escape Rope
-86 - Repel
-87 - ????????
-88 - ????????
-89 - ????????
-90 - ????????
-91 - ????????
-92 - ????????
-93 - Sun Stone
-94 - Moon Stone
-95 - Fire Stone
-96 - Thunderstone
-97 - Water Stone
-98 - Leaf Stone
-99 - ????????
-100 - ????????
-101 - ????????
-102 - ????????
-103 - Tinymushroom
-104 - Big Mushroom
-105 - ????????
-106 - Pearl
-107 - Big Pearl
-108 - Stardust
-109 - Star Piece
-110 - Nugget
-111 - Heart Scale
-112 - ????????
-113 - ????????
-114 - ????????
-115 - ????????
-116 - ????????
-117 - ????????
-118 - ????????
-119 - ????????
-120 - ????????
-121 - Orange Mail
-122 - Harbor Mail
-123 - Glitter Mail
-124 - Mech Mail
-125 - Wood Mail
-126 - Wave Mail
-127 - Bead Mail
-128 - Shadow Mail
-129 - Tropid Mail
-130 - Dream Mail
-131 - Fab Mail
-132 - Retro Mail
-133 - Cheri Berry
-134 - Chesto Berry
-135 - Pecha Berry
-136 - Rawst Berry
-137 - Aspear Berry
-138 - Leppa Berry
-139 - Orang Berry
-140 - Persim Berry
-141 - Lum Berry
-142 - Sitrus Berry
-143 - Figy Berry
-144 - Wiki Berry
-145 - Mago Berry
-146 - Aguav Berry
-147 - Iapapa Berry
-148 - Razz Berry
-149 - Bluk Berry
-150 - Nanab Berry
-151 - Wepear Berry
-152 - Pinap
-153 - Pomeg
-154 - Kelpsy
-155 - Qualot
-156 - Hondew
-157 - Grepa
-158 - Tamato
-159 - Cornn
-160 - Magost
-161 - Rabuta
-162 - Nomel
-163 - Spelon
-164 - Pamtre
-165 - Watmel
-166 - Durin
-167 - Belue
-168 - Liechi
-169 - Ganlon
-170 - Salac
-171 - Petaya
-172 - Apicot
-173 - Lansat
-174 - Starf
-175 - Enigma
-176 - ????????
-177 - ????????
-178 - ????????
-179 - Brightpowder
-180 - White Herb
-181 - Macho Brace
-182 - Exp. Share
-183 - Quick Claw
-184 - Soothe Bell
-185 - Mental Herb
-186 - Choice Band
-187 - King's Rock
-188 - Silverpowder
-189 - Amulet Coin
-190 - Cleanse Tag
-191 - Soul Dew
-192 - Deep Sea Tooth
-193 - Deep Sea Scale
-194 - Smoke Ball
-195 - Everstone
-196 - Focus Band
-197 - Lucky Egg
-198 - Scope Lens
-199 - Metal Coat
-200 - Leftovers
-201 - Dragon Scale
-202 - Light Ball
-203 - Soft Sand
-204 - Hard Stone
-205 - Miracle Seed
-206 - BlackGlasses
-207 - Black Belt
-208 - Magnet
-209 - Mystic Water
-210 - Sharp Beak
-213 - Poison Barb
-214 - Never Melt Ice
-215 - Spell Tag
-216 - Twisted Spoon
-217 - Charcoal
-218 - Dragon Fang
-219 - Silk Scarf
-220 - Up-Grade
-221 - Shell Bell
-222 - Sea Incense
-223 - Lax Incense
-224 - Lucky Punch
-225 - Metal Powder
-226 - Thick Club
-227 - Stick
-<end>
