--- data/pokeballs.txt
@@ -1,16 +1,0 @@
-0 - Pokeball
-1 - Master Ball
-2 - Ultra Ball
-3 - Great Ball
-4 - Poke Ball?
-5 - Safari Ball
-6 - Net Ball?
-7 - Dive Ball
-8 - Nest Ball
-9 - Repeat Ball
-10 - Timer Ball
-11 - Luxury Ball
-12 - Premiere Ball
-13 - Poke Ball?
-14 - Poke Ball?
-15 - Poke Ball
