+++ include/CMakeLists.txt
@@ -1,0 +1,1 @@
+ADD_SUBDIRECTORY(pokehack)
