+++ lib/CMakeLists.txt
@@ -1,0 +1,10 @@
+INCLUDE_DIRECTORIES(${POKEHACK_SOURCE_DIR}/include)
+
+FILE(GLOB pokehack_sources RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} "*.c*")
+ADD_LIBRARY(pokehack SHARED ${pokehack_sources})
+SET_TARGET_PROPERTIES(pokehack PROPERTIES DEFINE_SYMBOL "POKEHACK_DLL_EXPORTS")
+INSTALL(TARGETS pokehack
+        LIBRARY DESTINATION ${LIBRARY_DIR} COMPONENT Libraries #.so
+        ARCHIVE DESTINATION ${LIBRARY_DIR} COMPONENT Libraries #.lib
+        RUNTIME DESTINATION ${RUNTIME_DIR} COMPONENT Runtime #.dll
+)
