--- src/pokehack.pro
@@ -1,26 +1,0 @@
-SOURCES += main.cpp
-SOURCES += SaveParser.cpp
-SOURCES += window.cpp
-SOURCES += AttackComboBox.cpp
-SOURCES += ItemComboBox.cpp
-SOURCES += SpeciesComboBox.cpp
-SOURCES += PokeballComboBox.cpp
-SOURCES += StatusComboBox.cpp
-SOURCES += GameComboBox.cpp
-SOURCES += pokemon_species.c
-SOURCES += items.c
-SOURCES += misc.c
-SOURCES += attacks.c
-
-
-HEADERS += window.h
-HEADERS += ../include/pokestructs.h
-HEADERS += SaveParser.h
-HEADERS += AttackComboBox.h
-HEADERS += ItemComboBox.h
-HEADERS += SpeciesComboBox.h
-HEADERS += PokeballComboBox.h
-HEADERS += StatusComboBox.h
-HEADERS += GameComboBox.h
-
-CONFIG += qt
